refactor processing of samples
