+++ pkg/phlaredb/symdb/partition_memory_test.go
@@ -84,79 +84,6 @@ func Test_StacktraceAppender_shards(t *testing.T) {
 	})
 }
 
-func Test_StacktraceResolver_stacktraces_split(t *testing.T) {
-	type testCase struct {
-		description string
-		maxNodes    uint32
-		stacktraces []uint32
-		expected    []StacktracesRange
-	}
-
-	testCases := []testCase{
-		{
-			description: "no limit",
-			stacktraces: []uint32{234, 1234, 2345},
-			expected: []StacktracesRange{
-				{ids: []uint32{234, 1234, 2345}},
-			},
-		},
-		{
-			description: "one chunk",
-			maxNodes:    4,
-			stacktraces: []uint32{1, 2, 3},
-			expected: []StacktracesRange{
-				{m: 4, chunk: 0, ids: []uint32{1, 2, 3}},
-			},
-		},
-		{
-			description: "one chunk shifted",
-			maxNodes:    4,
-			stacktraces: []uint32{401, 402},
-			expected: []StacktracesRange{
-				{m: 4, chunk: 100, ids: []uint32{1, 2}},
-			},
-		},
-		{
-			description: "multiple shards",
-			maxNodes:    4,
-			stacktraces: []uint32{1, 2, 5, 7, 11, 13, 14, 15, 17, 41, 42, 43, 83, 85, 86},
-			//         : []uint32{1, 2, 1, 3,  3,  1,  2,  3,  1,  1,  2,  3,  3,  1,  2},
-			//         : []uint32{0, 0, 1, 1,  2,  3,  3,  3,  4, 10, 10, 10, 20, 21, 21},
-			expected: []StacktracesRange{
-				{m: 4, chunk: 0, ids: []uint32{1, 2}},
-				{m: 4, chunk: 1, ids: []uint32{1, 3}},
-				{m: 4, chunk: 2, ids: []uint32{3}},
-				{m: 4, chunk: 3, ids: []uint32{1, 2, 3}},
-				{m: 4, chunk: 4, ids: []uint32{1}},
-				{m: 4, chunk: 10, ids: []uint32{1, 2, 3}},
-				{m: 4, chunk: 20, ids: []uint32{3}},
-				{m: 4, chunk: 21, ids: []uint32{1, 2}},
-			},
-		},
-		{
-			description: "multiple shards exact",
-			maxNodes:    4,
-			stacktraces: []uint32{1, 2, 5, 7, 11, 13, 14, 15, 17, 41, 42, 43, 83, 85, 86, 87},
-			expected: []StacktracesRange{
-				{m: 4, chunk: 0, ids: []uint32{1, 2}},
-				{m: 4, chunk: 1, ids: []uint32{1, 3}},
-				{m: 4, chunk: 2, ids: []uint32{3}},
-				{m: 4, chunk: 3, ids: []uint32{1, 2, 3}},
-				{m: 4, chunk: 4, ids: []uint32{1}},
-				{m: 4, chunk: 10, ids: []uint32{1, 2, 3}},
-				{m: 4, chunk: 20, ids: []uint32{3}},
-				{m: 4, chunk: 21, ids: []uint32{1, 2, 3}},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.description, func(t *testing.T) {
-			assert.Equal(t, tc.expected, SplitStacktraces(tc.stacktraces, tc.maxNodes))
-		})
-	}
-}
-
 func Test_Stacktrace_append_existing(t *testing.T) {
 	db := NewSymDB(new(Config))
 	w := db.PartitionWriter(0)
