+++ pkg/phlaredb/symdb/stacktrace_selection_test.go
